V10_1-HTCONDOR-1337-no-more-floats

+++ src/condor_shadow.V6.1/shadow.cpp
@@ -228,7 +228,7 @@ void
 UniShadow::emailTerminateEvent( int exitReason, update_style_t kind )
 {
 	Email mailer;
-	float recvd_bytes = 0, sent_bytes = 0;
+	double recvd_bytes = 0, sent_bytes = 0;
 
 	if (kind == US_TERMINATE_PENDING) {
 		/* I don't have a remote resource, so get the values directly from
