+++ src/condor_tools/give_interactive.cpp
@@ -320,7 +320,7 @@ fetchSubmittorPrios()
 	ClassAd	al;
 	char  	attrName[32], attrPrio[32];
   	char  	name[128];
-  	float 	sub_priority;
+  	double 	sub_priority;
 	int		i = 1;
 
 		// Minor hack, if we're talking to a remote pool, assume the
